refactor python bindings
