+++ apis/python/src/tiledbvcf/binding/libtiledbvcf.cc
@@ -1,62 +1,52 @@
+/**
+ * @section LICENSE
+ *
+ * The MIT License
+ *
+ * @copyright Copyright (c) 2019-2024 TileDB, Inc.
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+
 #include <pybind11/pybind11.h>
 #include <pybind11/stl_bind.h>
 #include <tiledbvcf/tiledbvcf.h>
 
-#include <map>
-#include <stdexcept>
+// C++ API bindings
+#include "sample_stats.h"
 
+// C API bindings
 #include "reader.h"
 #include "writer.h"
 
-#include "stats/arrow_adapter.h"
-#include "stats/sample_stats.h"
+namespace tiledbvcfpy {
 
 namespace py = pybind11;
-using namespace py::literals;
-using namespace tiledbvcfpy;
 using namespace tiledb::vcf;
 
-/**
- * @brief Convert ArrayBuffers to Arrow table.
- *
- * @param cbs ArrayBuffers
- * @return py::object
- */
-py::object _buffer_to_table(std::shared_ptr<ArrayBuffers> buffers) {
-  auto pa = py::module::import("pyarrow");
-  auto pa_table_from_arrays = pa.attr("Table").attr("from_arrays");
-  auto pa_array_import = pa.attr("Array").attr("_import_from_c");
-  auto pa_schema_import = pa.attr("Schema").attr("_import_from_c");
-
-  py::list array_list;
-  py::list names;
-
-  // Process each column buffer in the expected order
-  for (auto& name : buffers->names()) {
-    printf("convert buffer: %s\n", name.c_str());
-    auto column = buffers->at(name);
-    auto [pa_array, pa_schema] = ArrowAdapter::to_arrow(column);
-    auto array = pa_array_import(
-        py::capsule(pa_array.get()), py::capsule(pa_schema.get()));
-    array_list.append(array);
-    names.append(name);
-  }
-
-  return pa_table_from_arrays(array_list, names);
-}
-
-std::optional<py::object> to_table(
-    std::optional<std::shared_ptr<ArrayBuffers>> buffers) {
-  if (buffers.has_value()) {
-    return _buffer_to_table(*buffers);
-  }
-
-  return std::nullopt;
-}
-
 PYBIND11_MODULE(libtiledbvcf, m) {
-  m.doc() = "Python binding of the TileDB-VCF library C API.";
+  m.doc() = "Python binding for TileDB-VCF.";
 
+  // C++ API bindings
+  load_sample_stats(m);
+
+  // C API bindings
   m.def("config_logging", &config_logging);
 
   py::class_<Reader>(m, "Reader")
@@ -162,17 +152,6 @@ PYBIND11_MODULE(libtiledbvcf, m) {
       .def("set_enable_sample_stats", &Writer::set_enable_sample_stats)
       .def("set_compress_sample_dim", &Writer::set_compress_sample_dim)
       .def("set_compression_level", &Writer::set_compression_level);
-
-  m.def(
-      "sample_qc",
-      [](const std::string& dataset_uri,
-         const std::vector<std::string>& samples,
-         const std::map<std::string, std::string>& config)
-          -> std::optional<py::object> {
-        auto buffers = SampleStats::sample_qc(dataset_uri, samples, config);
-        return to_table(buffers);
-      },
-      "dataset_uri"_a,
-      "samples"_a,
-      "config"_a);
 }
+
+}  // namespace tiledbvcfpy
