+++ apis/python/src/tiledbvcf/binding/vcf_arrow.h
@@ -1,6 +1,4 @@
 /**
- * @file   vcf_arrow.h
- *
  * @section LICENSE
  *
  * The MIT License
@@ -33,7 +31,7 @@
 #include <pybind11/stl.h>
 #include <tiledbvcf/tiledbvcf.h>
 
-#include "carrow.h"
+#include "stats/carrow.h"
 
 namespace py = pybind11;
